+++ src/tapis-app/Streams/Instruments/_Layout/Layout.tsx
