+++ src/tapis-app/Streams/Instruments/_Layout/index.ts
