+++ src/tapis-app/Streams/Instruments/_components/InstrumentsNav/InstrumentsNav.test.tsx
