+++ src/tapis-app/Streams/Instruments/_components/InstrumentsNav/index.ts
