+++ src/tapis-app/Streams/Instruments/_components/index.ts
