+++ src/tapis-app/Streams/Projects/_Layout/Layout.tsx
@@ -1,23 +1,8 @@
 import React from 'react';
-import {
-  PageLayout,
-  LayoutHeader
-} from 'tapis-ui/_common';
-
 import { Router } from '../_Router';
 
 const Layout: React.FC = () => {
-  const header = (
-    <LayoutHeader>
-      <div>Projects</div>
-    </LayoutHeader>
-  );
-
-  const body = (
-    <Router />
-  );
-
-  return <PageLayout top={header} left={body} />;
+  return <Router />;
 };
 
 export default Layout;
