+++ src/tapis-app/Streams/Projects/_Router/Router.tsx
@@ -7,14 +7,29 @@ import {
 } from 'react-router-dom';
 import Sites from "../../Sites";
 import { ProjectsNav } from '../_components';
+import {
+  PageLayout,
+  LayoutHeader
+} from 'tapis-ui/_common';
 
 const Router: React.FC = () => {
+
+  const header = (
+    <LayoutHeader>
+      <div>Projects</div>
+    </LayoutHeader>
+  );
+  
+  const body = (
+    <ProjectsNav />
+  );
+
   const { path } = useRouteMatch();
 
   return (
     <Switch>
       <Route path={`${path}`} exact>
-        <ProjectsNav />
+        <PageLayout top={header} left={body} />
       </Route>
 
       <Route
