+++ src/tapis-app/Streams/Sites/_Layout/Layout.tsx
@@ -1,24 +1,10 @@
 import React from 'react';
-import {
-  PageLayout,
-  LayoutHeader
-} from 'tapis-ui/_common';
 import { Router } from '../_Router';
 
 const Layout: React.FC<{ projectId: string }> = ({
   projectId
 }) => {
-  const header = (
-    <LayoutHeader>
-      <div>Sites</div>
-    </LayoutHeader>
-  );
-
-  const body = (
-    <Router projectId={projectId} />
-  );
-
-  return <PageLayout top={header} left={body} />;
+  return <Router projectId={projectId} />;
 };
 
 export default Layout;
