+++ src/tapis-app/Streams/_Layout/Layout.tsx
@@ -1,25 +1,10 @@
 import React from 'react';
-import {
-  PageLayout,
-  LayoutHeader,
-} from 'tapis-ui/_common';
 import Projects from "../Projects";
 
 const Layout: React.FC = () => {
-  const header = (
-    <LayoutHeader>
-      <div>TAPIS Streams</div>
-    </LayoutHeader>
-  );
-
-  const body = (
-    <Projects />
-  );
 
   return (
-    <>
-      <PageLayout top={header} left={body} />
-    </>
+    <Projects />
   );
 };
